+++ app/Http/Controllers/StructureController.php
@@ -6,45 +6,61 @@ use App\Models\Status;
 use App\Models\Structure;
 use App\Models\Tasks;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class StructureController extends Controller
 {
     public function getStructure(){
-        return Structure::whereNull('structure_id')
-            ->with('children')
-            ->get();
+
+        $userAuth = Auth::user();
+        if($userAuth->hasPermission('view-structure-page')) {
+            return Structure::whereNull('structure_id')
+                ->with('children')
+                ->get();
+        }
     }
 
     public function createCategory(Request $request){
-        Structure::create([
-            'name' => $request->name,
-            'structure_id' => $request->parent_id,
-        ]);
+        $userAuth = Auth::user();
+        if($userAuth->hasPermission('update-structure')) {
+            Structure::create([
+                'name' => $request->name,
+                'structure_id' => $request->parent_id,
+            ]);
+        }
     }
     public function updateCategory(Request $request){
-        $event = Structure::find($request->id);
-        $event->update([
-            'name' => $request->name,
-        ]);
-
+        $userAuth = Auth::user();
+        if($userAuth->hasPermission('update-structure')) {
+            $event = Structure::find($request->id);
+            $event->update([
+                'name' => $request->name,
+            ]);
+        }
     }
     public function deleteCategory($id){
-        Structure::where('id', $id)->delete();
+        $userAuth = Auth::user();
+        if($userAuth->hasPermission('update-structure')) {
+            Structure::where('id', $id)->delete();
+        }
     }
 
 
     public function tasksByStructure($id){
-        $structure = Structure::where('id', $id)->with('children')->first();
 
-        $ids = $this->getIds($structure->children);
-        array_push($ids, $structure->id);
+        $userAuth = Auth::user();
+        if($userAuth->hasPermission('view-structure-page')) {
+            $structure = Structure::where('id', $id)->with('children')->first();
 
-        $status = Status::with(['statusTasks' => function($query) use ($ids) {
-            $query->with('initiator', 'status', 'responsibles', 'priority')->whereIn('structure_id', $ids);
-        }])->get();
+            $ids = $this->getIds($structure->children);
+            array_push($ids, $structure->id);
 
-//        return Tasks::whereIn('structure_id', $ids)->get();
-        return $status;
+            $status = Status::with(['statusTasks' => function($query) use ($ids) {
+                $query->with('initiator', 'status', 'responsibles', 'priority')->whereIn('structure_id', $ids);
+            }])->get();
+
+            return $status;
+        }
     }
 
     public function getIds($items){
